requirements.txt
aniso8601==4.0.1
Click==7.0
cycler==0.10.0
Django==2.1.3
docutils==0.14
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lockfile==0.12.2
MarkupSafe==1.1.0
matplotlib==3.0.2
mpld3==0.3
numpy==1.15.4
picamera==1.13
pyparsing==2.3.0
python-daemon==2.2.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
SQLAlchemy==1.2.14
Werkzeug==0.14.1
